Use new firewall scripts
The original 'command' backend uses the shell extensively to construct
and execute firewall commands. Instead of stuffing complex constructs to
fit the limits of the backend, delegate firewall actions to a script.

On startup, the new script initializes the firewall. It reads commands
from standard input, one per line, that either block, release, or flush
addresses from the firewall. SSHGuard controls this script using a pipe.
This will ease privilege separation in the future.

The current implementation is not quite portable because it uses
popen(), and cannot yet execute external commands. Only the pf, ipfw,
and null backends are currently implemented.